+++ src/main/java/com/github/lipinskipawel/board/ai/BoardEvaluator.java
@@ -1,9 +1,10 @@
 package com.github.lipinskipawel.board.ai;
 
 import com.github.lipinskipawel.board.engine.Board;
+import com.github.lipinskipawel.board.engine.Player;
 
 public interface BoardEvaluator {
 
-    double evaluate(Board board);
+    double evaluate(Board<Player> board);
 
 }