+++ src/main/java/com/github/lipinskipawel/board/ai/MoveStrategy.java
@@ -3,6 +3,7 @@ package com.github.lipinskipawel.board.ai;
 import com.github.lipinskipawel.board.ai.bruteforce.DefaultMoveStrategyBuilder;
 import com.github.lipinskipawel.board.engine.Board;
 import com.github.lipinskipawel.board.engine.Move;
+import com.github.lipinskipawel.board.engine.Player;
 
 /**
  * The underlying implementations MUST guarantee thread safety.
@@ -16,7 +17,7 @@ public interface MoveStrategy {
      * @param board to search best move on
      * @return best move
      */
-    Move searchForTheBestMove(Board board);
+    Move searchForTheBestMove(Board<Player> board);
 
     /**
      * This method provides default {@link MoveStrategy} builder. Builder already